refactor(taiko-client-rs): remove dead embedded-node stack and dead public surface
Delete code verified to have zero callers across the workspace:

- preconfirmation-driver: the superseded channel-based integration stack
  (PreconfirmationDriverNode, DriverChannels, EmbeddedDriverClient,
  NodeRpcApiImpl, node_integration tests) — production uses
  PreconfirmationDriverRunner with EventSyncerDriverClient; the channel
  protocol had no receiver in the driver crate. ContractInboxReader moves
  to driver_interface/inbox.rs and the shared RPC helpers to
  rpc/api_helpers.rs. Also drop the unused EventLoop restart machinery
  (run_with_retry, rebuild_after_failure, RetryBackoff, handler setters)
  and the never-constructed Storage/ChannelClosed/ValidationErrorCode::Other
  error variants with their unreachable mapping arms.
- driver: drop the unused preconfirmation_sender/Driver::config/
  PreconfPayload::into_payload methods, the protocol::signer re-export,
  the dead sync Result alias, and never-constructed error variants
  (DriverError::Io, DerivationError::Validation,
  EngineSubmissionError::MissingParent).
- protocol: drop FixedKSigner's unused alloy Signer/SignerSync impls and
  chain_id field (all callers use sign_with_predefined_k directly) and
  never-constructed LookaheadError::{Lookahead, ReorgDetected} and
  ProtocolError::{Rlp, Other} variants.
- rpc: drop the uncalled last_certain_l1_origin_by_batch_id engine method
  and a From impl covering a transport-error type nobody constructs.
- whitelist-preconfirmation-driver: retarget the cache-import deferral
  test from the removed MissingParent variant to MissingPayloadId.
