+++ packages/taiko-client-rs/crates/driver/src/lib.rs
@@ -29,6 +29,3 @@ pub use driver::Driver;
 pub use error::{DriverError, map_driver_error};
 pub use production::PreconfPayload;
 pub use sync::{ConfirmedSyncSnapshot, SyncPipeline, SyncStage, event::EventSyncer};
-
-// Re-export signer from protocol crate for backward compatibility
-pub use protocol::signer;
