+++ packages/taiko-client-rs/crates/preconfirmation-driver/Cargo.toml
@@ -63,7 +63,3 @@ path = "tests/catchup_e2e.rs"
 [[test]]
 name = "head_gossip_e2e"
 path = "tests/head_gossip_e2e.rs"
-
-[[test]]
-name = "node_integration"
-path = "tests/node_integration.rs"
