+++ packages/taiko-client-rs/crates/preconfirmation-driver/src/rpc/api_helpers.rs
@@ -1,6 +1,4 @@
-//! RPC API implementation backed by the preconfirmation driver node state.
-
-use std::sync::Arc;
+//! Shared helpers backing the preconfirmation RPC API implementation.
 
 use alloy_primitives::{B256, U256};
 use preconfirmation_net::NetworkCommand;
@@ -10,7 +8,7 @@ use preconfirmation_types::{
 };
 use protocol::codec::ZlibTxListCodec;
 use ssz_rs::Deserialize;
-use tokio::sync::{mpsc, watch};
+use tokio::sync::mpsc;
 use tracing::{info, warn};
 
 use crate::{
@@ -18,67 +16,13 @@ use crate::{
     driver_interface::{DriverClient, InboxReader, PreconfirmationInput},
     error::{PreconfirmationClientError, ValidationErrorCode},
     metrics::PreconfirmationClientMetrics,
-    rpc::{NodeStatus, PreconfRpcApi, PreconfSlotInfo, PublishBlockRequest, PublishBlockResponse},
+    rpc::{NodeStatus, PublishBlockRequest, PublishBlockResponse},
     validation::{is_eop_only, validate_commitment_with_signer, validate_lookahead},
 };
 
 /// Increment the counter tracking rejected preconfirmation inputs.
 fn record_validation_failure() {
     PreconfirmationClientMetrics::validation_failures_total().inc();
-}
-
-/// Internal RPC API implementation backed by the preconfirmation driver node state.
-pub(crate) struct NodeRpcApiImpl<I: InboxReader, D: DriverClient> {
-    /// Command tx for issuing commands to the P2P network layer.
-    pub(crate) command_tx: mpsc::Sender<NetworkCommand>,
-    /// Watch receiver for the preconfirmation tip.
-    pub(crate) preconf_tip_rx: watch::Receiver<U256>,
-    /// Inbox reader for checking L1 sync state.
-    pub(crate) inbox_reader: I,
-    /// Lookahead resolver for slot info by timestamp.
-    pub(crate) lookahead_resolver:
-        Arc<dyn protocol::preconfirmation::PreconfSignerResolver + Send + Sync>,
-    /// Driver client for submitting preconfirmation inputs.
-    pub(crate) driver: Arc<D>,
-    /// Txlist codec for decompression.
-    pub(crate) codec: Arc<ZlibTxListCodec>,
-    /// Expected slasher address for commitment validation.
-    pub(crate) expected_slasher: Option<Bytes20>,
-    /// Local peer ID string used in status responses.
-    pub(crate) local_peer_id: String,
-}
-
-#[async_trait::async_trait]
-impl<I: InboxReader + 'static, D: DriverClient + 'static> PreconfRpcApi for NodeRpcApiImpl<I, D> {
-    async fn publish_block(&self, request: PublishBlockRequest) -> Result<PublishBlockResponse> {
-        publish_block_impl(
-            &self.command_tx,
-            self.driver.as_ref(),
-            &self.codec,
-            self.expected_slasher.as_ref(),
-            self.lookahead_resolver.as_ref(),
-            request,
-        )
-        .await
-    }
-
-    async fn get_status(&self) -> Result<NodeStatus> {
-        let preconf_tip = *self.preconf_tip_rx.borrow();
-        build_node_status(&self.command_tx, &self.inbox_reader, preconf_tip, &self.local_peer_id)
-            .await
-    }
-
-    async fn preconf_tip(&self) -> Result<U256> {
-        Ok(*self.preconf_tip_rx.borrow())
-    }
-
-    async fn get_preconf_slot_info(&self, timestamp: U256) -> Result<PreconfSlotInfo> {
-        self.lookahead_resolver
-            .slot_info_for_timestamp(timestamp)
-            .await
-            .map(PreconfSlotInfo::from)
-            .map_err(Into::into)
-    }
 }
 
 /// Query the P2P network for the current peer count.
@@ -284,7 +228,6 @@ pub(crate) async fn build_node_status<I: InboxReader>(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::sync::atomic::{AtomicU64, Ordering};
 
     /// Mock lookahead resolver for testing.
     struct MockLookaheadResolver;
@@ -308,53 +251,6 @@ mod tests {
         }
     }
 
-    /// Mock inbox reader for testing.
-    #[derive(Clone)]
-    struct MockInboxReader {
-        next_proposal_id: std::sync::Arc<AtomicU64>,
-        target_block: std::sync::Arc<AtomicU64>,
-        head_l1_origin_block_id: std::sync::Arc<AtomicU64>,
-    }
-
-    const NONE_SENTINEL: u64 = u64::MAX;
-
-    impl MockInboxReader {
-        fn new(
-            next_proposal_id: u64,
-            target_block: Option<u64>,
-            head_l1_origin: Option<u64>,
-        ) -> Self {
-            Self {
-                next_proposal_id: std::sync::Arc::new(AtomicU64::new(next_proposal_id)),
-                target_block: std::sync::Arc::new(AtomicU64::new(
-                    target_block.unwrap_or(NONE_SENTINEL),
-                )),
-                head_l1_origin_block_id: std::sync::Arc::new(AtomicU64::new(
-                    head_l1_origin.unwrap_or(NONE_SENTINEL),
-                )),
-            }
-        }
-
-        fn read_optional(value: u64) -> Option<u64> {
-            (value != NONE_SENTINEL).then_some(value)
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl InboxReader for MockInboxReader {
-        async fn get_next_proposal_id(&self) -> Result<u64> {
-            Ok(self.next_proposal_id.load(Ordering::SeqCst))
-        }
-
-        async fn get_last_block_id_by_batch_id(&self, _proposal_id: u64) -> Result<Option<u64>> {
-            Ok(Self::read_optional(self.target_block.load(Ordering::SeqCst)))
-        }
-
-        async fn get_head_l1_origin_block_id(&self) -> Result<Option<u64>> {
-            Ok(Self::read_optional(self.head_l1_origin_block_id.load(Ordering::SeqCst)))
-        }
-    }
-
     /// Stub driver that accepts everything and keeps tip at zero.
     struct StubDriver;
     #[async_trait::async_trait]
@@ -371,59 +267,6 @@ mod tests {
         async fn preconf_tip(&self) -> Result<U256> {
             Ok(U256::ZERO)
         }
-    }
-
-    #[tokio::test]
-    async fn test_node_status_includes_peer_id() {
-        let (_preconf_tip_tx, preconf_tip_rx) = watch::channel(U256::from(100));
-        let (command_tx, mut command_rx) = mpsc::channel::<NetworkCommand>(16);
-        let local_peer_id = "test-peer-id".to_string();
-
-        let api = NodeRpcApiImpl {
-            command_tx,
-            preconf_tip_rx,
-            inbox_reader: MockInboxReader::new(43, Some(120), Some(120)),
-            lookahead_resolver: Arc::new(MockLookaheadResolver),
-            driver: Arc::new(StubDriver),
-            codec: Arc::new(ZlibTxListCodec::new(preconfirmation_types::MAX_TXLIST_BYTES)),
-            expected_slasher: None,
-            local_peer_id: local_peer_id.clone(),
-        };
-
-        tokio::spawn(async move {
-            if let Some(NetworkCommand::GetPeerCount { respond_to }) = command_rx.recv().await {
-                let _ = respond_to.send(5);
-            }
-        });
-
-        let status = api.get_status().await.unwrap();
-        assert_eq!(status.peer_id, local_peer_id);
-        assert_eq!(status.peer_count, 5);
-        assert!(status.is_synced_with_inbox);
-        assert_eq!(status.event_sync_tip, Some(U256::from(120)));
-    }
-
-    #[tokio::test]
-    async fn test_get_preconf_slot_info_returns_resolver_output() {
-        let (_preconf_tip_tx, preconf_tip_rx) = watch::channel(U256::ZERO);
-        let (command_tx, mut command_rx) = mpsc::channel::<NetworkCommand>(16);
-
-        let api = NodeRpcApiImpl {
-            command_tx,
-            preconf_tip_rx,
-            inbox_reader: MockInboxReader::new(0, None, None),
-            lookahead_resolver: Arc::new(MockLookaheadResolver),
-            driver: Arc::new(StubDriver),
-            codec: Arc::new(ZlibTxListCodec::new(preconfirmation_types::MAX_TXLIST_BYTES)),
-            expected_slasher: None,
-            local_peer_id: "test".to_string(),
-        };
-
-        tokio::spawn(async move { while command_rx.recv().await.is_some() {} });
-
-        let slot_info = api.get_preconf_slot_info(U256::from(500)).await.unwrap();
-        assert_eq!(slot_info.signer, alloy_primitives::Address::repeat_byte(0x11));
-        assert_eq!(slot_info.submission_window_end, U256::from(2000));
     }
 
     #[tokio::test]
